+++ src/index.test.ts
@@ -179,7 +179,7 @@ describe("restoreSession", () => {
     );
 
     await expect(AuthN.restoreSession()).rejects.toEqual([
-      { message: "Unauthorized" },
+      { message: "401" },
     ]);
     expect(AuthN.session()).toBeFalsy();
   });
@@ -209,7 +209,7 @@ describe("restoreSession", () => {
     );
 
     await expect(AuthN.restoreSession()).rejects.toEqual([
-      { message: "Unauthorized" },
+      { message: "401" },
     ]);
     expect(AuthN.session()).toBeFalsy();
   });