+++ src/verbs.ts
@@ -38,8 +38,8 @@ function jhr<T>(sender: (xhr: XMLHttpRequest) => void): Promise<T> {
         if ('errors' in data) {
           reject(data.errors)
         } else if (xhr.status > 400) {
-          // statusText may be missing in HTTP/2
-          reject([{message: xhr.statusText || xhr.status.toString()}])
+          // statusText may be missing in HTTP/2. only the status number is reliable.
+          reject([{message: xhr.status.toString()}])
         } else {
           fulfill(data.result)
         }